requirements.txt
python-telegram-bot==20.7
requests==2.31.0
fastapi==0.115.0
uvicorn==0.23.2